+++ backend/src/models/govt_agency.ts
@@ -1,14 +1,5 @@
 import { Schema, model } from 'mongoose';
-import { Coordinate } from '../utils/coordinates';
-
-type GovtAgency = {
-  name: string;
-  description?: string;
-  email: string;
-  phone: string[];
-  location: Coordinate;
-  address: string;
-};
+import { GovtAgency } from '../types/Schema'
 
 const GovtAgencySchema = new Schema<GovtAgency>({
   name: {
