+++ backend/src/models/rescue_agency.ts
@@ -1,17 +1,5 @@
 import { Schema, model } from 'mongoose';
-import { Coordinate } from '../utils/coordinates';
-
-type RescueAgency = {
-  name: string;
-  description?: string;
-  email: string;
-  phone: string[];
-  location: Coordinate;
-  address: string;
-  type: string;
-  created_at: Date;
-  updated_at: Date;
-};
+import { RescueAgency } from '../types/schema';
 
 const RescueAgencySchema = new Schema<RescueAgency>(
   {
