+++ backend/src/models/resource.ts
@@ -1,14 +1,6 @@
-import { Schema, model, Types } from 'mongoose';
+import { Schema, model } from 'mongoose';
+import { Resource } from '../types/schema';
 
-type Resource = {
-  agency_id: Types.ObjectId;
-  type: string;
-  name: string;
-  quantity: number;
-  unit: string;
-  created_at: Date;
-  updated_at: Date;
-};
 
 const ResourceSchema = new Schema<Resource>(
   {
