--- backend/src/utils/coordinates.ts
@@ -1,4 +1,0 @@
-export type Coordinate = {
-  latitude: number;
-  longitude: number;
-};
